(Obj-C wrapper) Fixes all but two warnings - will need to replace amqp_set_sockfd parts in AMQPConnection with SSL equivalent

+++ librabbitmq-ios/AMQPConsumer.m
@@ -95,7 +95,7 @@
 		
 		properties = (amqp_basic_properties_t*)frame.payload.properties.decoded;
 		
-		bodySize = frame.payload.properties.body_size;
+		bodySize = (size_t)frame.payload.properties.body_size;
 		receivedBytes = 0;
 		body = amqp_bytes_malloc(bodySize);
